add header file to population and consoleInterface

+++ main.cpp
@@ -1,4 +1,7 @@
-#include "view/consoleInterface.cpp"
+#include <cstdlib>
+#include <string>
+
+#include "view/consoleInterface.hpp"
 
 int main(int argc, char const *argv[]) {
   view::ConsoleInterface consoleInterface{};
@@ -6,8 +9,8 @@ int main(int argc, char const *argv[]) {
     const int S = atoi(argv[1]);
     const int I = atoi(argv[2]);
     const int R = atoi(argv[3]);
-    const float beta = std::stof(argv[4]);
-    const float gamma = std::stof(argv[5]);
+    const double beta = std::stof(argv[4]);
+    const double gamma = std::stof(argv[5]);
     const int t = atoi(argv[6]);
     consoleInterface.start(S, I, R, beta, gamma, t);
   } else {
